+++ Source/main.cpp
@@ -4,7 +4,7 @@
 /// @brief      Holds all the main functions.
 /// @author     GamerMan7799
 /// @author     xPUREx
-/// @version    4.2.0-Beta.5
+/// @version    4.2.0-Beta.6
 /// @date       2015
 /// @copyright  Public Domain Unlicense.
 /////////////////////////////////////////////////
@@ -17,9 +17,6 @@
 #include "entity.h"
 #include "globals.h"
 #include "screen.h"
-/*****************************************************************************/
-//General Todos
-/** \todo (xPUREx#2#): Add a short message the varies with the Software status (ex warning if using a beta or alpha) */
 /*****************************************************************************/
 /** This is the namespace that is meant to hold things that need to be referenced between
     different files */
@@ -67,6 +64,23 @@ int main(int argc, char *argv[]) {
     ///                  has completed its run.
     ///
     /////////////////////////////////////////////////
+
+    if ( DEFINED_VER_STATUS_SHORT == "a" ) {
+        printf("This version is an Alpha, meaning that we cannot be sure that it works correctly. ");
+        printf("Any features that you see may or may not be changed/removed later in the developmental cycle.\n\n");
+    } else if ( DEFINED_VER_STATUS_SHORT == "b" ) {
+        printf("This version is a Beta, meaning that while it should work very closely to how we intend it to work, ");
+        printf("there may still be some bugs that can cause the program to crash.\n\n");
+    } else if ( DEFINED_VER_STATUS_SHORT == "rc" ) {
+        printf("This version is a Release Candidate, meaning that it is only a few items away from being an Official ");
+        printf("Release. There shouldn't be very many bugs but let us know if you find any.\n\n");
+    } else if ( DEFINED_VER_STATUS_SHORT == "r" ) {
+        printf("This version is the official release for the game. There should not be any bugs, so if you find any ");
+        printf("please let us know right away so we can work to fix them.\n\n");
+    } else {
+        printf("This version is not properly labeled. There is a mistake in the version.h file.\n\n");
+    }
+
 
 	Global::Cnfg.Check(); //Load the config file's values
 	if (Global::blnError) {printf("\nThere was an error!\n"); return 1;}